+++ kernel/src/main/java/com/flyemu/share/dto/sales/SalesOrderDto.java
@@ -74,7 +74,7 @@ public class SalesOrderDto {
     private List<String> outOrderNoList;
 
     @Comment("商品数量")
-    private AtomicReference<Double> totalQuantity;
+    private AtomicReference<BigDecimal> totalQuantity;
 
     @Comment("出库单状态 0初始化 1部分出库 2全部出库")
     private Integer status;
